+++ src/main/kotlin/org/jetbrains/intellij/IntelliJPluginConstants.kt
@@ -1,6 +1,7 @@
 package org.jetbrains.intellij
 
 object IntelliJPluginConstants {
+    const val NAME = "Gradle IntelliJ Plugin"
     const val GROUP_NAME = "intellij"
     const val EXTENSION_NAME = "intellij"
     const val DEFAULT_SANDBOX = "idea-sandbox"