+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## next
 
+- Provide build information within the `META-INF/MANIFEST.MF` file
 - Resolve EAP version of the Java compiler for `LATEST-EAP-SNAPSHOT`
 - Allow for using `dcevm`, `fd`, and `nomod` variants of JBR [#818](../../issues/818)